chore: update package names
