+++ RemootioAndroidAuto/automotive/src/main/java/chamilton0/remootioandroidws/RemootioDeviceScreen.kt
@@ -1,25 +1,25 @@
-package com.example.remootioandroidws
+package chamilton0.remootioandroidws
 
 import androidx.car.app.CarContext
 import androidx.car.app.CarToast
 import androidx.car.app.Screen
 import androidx.car.app.model.*
 
-class RemootioScreen(carContext: CarContext?) : Screen(carContext!!) {
+class RemootioDeviceScreen(carContext: CarContext?) : Screen(carContext!!) {
     override fun onGetTemplate(): Template {
         val templateBuilder = ListTemplate.Builder()
         val radioList = ItemList.Builder()
-            .addItem(Row.Builder().setTitle("Garage Door").build())
-            .addItem(Row.Builder().setTitle("Gate").build())
+            .addItem(Row.Builder().setTitle("test").build())
             .setOnSelectedListener { index: Int -> onSelected(index) }
             .build()
         templateBuilder.addSectionedList(
-            SectionedItemList.create(radioList, "Remootio Devices")
+            SectionedItemList.create(radioList, "Test")
         )
-        return templateBuilder.setTitle("Remootio").build()
+        return templateBuilder.setTitle("2").setHeaderAction(Action.BACK).build()
     }
 
     private fun onSelected(index: Int) {
-        this.screenManager.push(RemootioDeviceScreen(carContext))
+        CarToast.makeText(carContext, "Changed selection to index: $index", CarToast.LENGTH_LONG)
+            .show()
     }
 }
