+++ RemootioAndroidAuto/automotive/src/main/java/chamilton0/remootioandroidws/RemootioService.kt
@@ -1,4 +1,4 @@
-package com.example.remootioandroidws
+package chamilton0.remootioandroidws
 
 import android.content.Intent
 import android.content.pm.ApplicationInfo
