--- RemootioAndroidAuto/automotive/src/main/java/com/example/remootioandroidws/RemootioDeviceScreen.kt
@@ -1,25 +1,0 @@
-package com.example.remootioandroidws
-
-import androidx.car.app.CarContext
-import androidx.car.app.CarToast
-import androidx.car.app.Screen
-import androidx.car.app.model.*
-
-class RemootioDeviceScreen(carContext: CarContext?) : Screen(carContext!!) {
-    override fun onGetTemplate(): Template {
-        val templateBuilder = ListTemplate.Builder()
-        val radioList = ItemList.Builder()
-            .addItem(Row.Builder().setTitle("test").build())
-            .setOnSelectedListener { index: Int -> onSelected(index) }
-            .build()
-        templateBuilder.addSectionedList(
-            SectionedItemList.create(radioList, "Test")
-        )
-        return templateBuilder.setTitle("2").setHeaderAction(Action.BACK).build()
-    }
-
-    private fun onSelected(index: Int) {
-        CarToast.makeText(carContext, "Changed selection to index: $index", CarToast.LENGTH_LONG)
-            .show()
-    }
-}
