+++ RemootioAndroidAuto/automotive/src/test/java/chamilton0/remootioandroidws/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.remootioandroidws;
+package chamilton0.remootioandroidws;
 
 import org.junit.Test;
